requirements.txt
litellm>=1.0.0
anthropic>=0.40.0
pyyaml>=6.0
streamlit>=1.32.0
pandas>=2.0.0
